+++ app/src/main/java/dev/nadeldrucker/trafficswipe/data/publicTransport/apis/generic/Entrypoint.java
@@ -1,0 +1,15 @@
+package dev.nadeldrucker.trafficswipe.data.publicTransport.apis.generic;
+
+import com.android.volley.RequestQueue;
+
+/**
+ * Entrypoint for api, exposing operations.
+ */
+public abstract class Entrypoint implements Capabilities {
+
+    protected RequestQueue queue;
+    public Entrypoint(RequestQueue queue){
+        this.queue = queue;
+    }
+
+}
