+++ app/src/main/java/dev/nadeldrucker/trafficswipe/data/publicTransport/model/data/Location.java
@@ -1,0 +1,31 @@
+package dev.nadeldrucker.trafficswipe.data.publicTransport.model.data;
+
+/**
+ * Location specified by latitude and longitude.
+ */
+public class Location {
+
+    private double latitude;
+    private double longitude;
+
+    public Location(double latitude, double longitude) {
+        this.latitude = latitude;
+        this.longitude = longitude;
+    }
+
+    public double getLatitude() {
+        return latitude;
+    }
+
+    public void setLatitude(double latitude) {
+        this.latitude = latitude;
+    }
+
+    public double getLongitude() {
+        return longitude;
+    }
+
+    public void setLongitude(double longitude) {
+        this.longitude = longitude;
+    }
+}
