--- app/src/main/java/dev/nadeldrucker/trafficswipe/data/transport/model/connection/RequestException.java
@@ -1,12 +1,0 @@
-package dev.nadeldrucker.trafficswipe.data.transport.model.connection;
-
-public class RequestException extends Exception {
-
-    public RequestException(String message) {
-        super(message);
-    }
-
-    public RequestException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
