--- app/src/main/java/dev/nadeldrucker/trafficswipe/data/transport/model/data/TransportEntity.java
@@ -1,21 +1,0 @@
-package dev.nadeldrucker.trafficswipe.data.transport.model.data;
-
-import com.android.volley.RequestQueue;
-
-import java.io.Serializable;
-
-/**
- * Superclass of all public transport data entities.
- */
-public abstract class TransportEntity implements Serializable {
-
-    private RequestQueue queue;
-
-    public TransportEntity(RequestQueue queue) {
-        this.queue = queue;
-    }
-
-    public RequestQueue getQueue() {
-        return queue;
-    }
-}
