--- app/src/main/java/dev/nadeldrucker/trafficswipe/data/transport/model/data/vehicle/Tram.java
@@ -1,28 +1,0 @@
-package dev.nadeldrucker.trafficswipe.data.transport.model.data.vehicle;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import com.android.volley.RequestQueue;
-import dev.nadeldrucker.trafficswipe.R;
-import dev.nadeldrucker.trafficswipe.data.transport.model.data.DepartureTime;
-import dev.nadeldrucker.trafficswipe.data.transport.model.data.Station;
-
-import java.util.Map;
-
-public class Tram extends Vehicle {
-
-    /**
-     * @param lineId            non-unique identifier for the specified transportation line
-     * @param direction         direction of vehicle as seen on the display
-     * @param entityId          unique object identifier for later use, if provided by the api
-     * @param targetDestination destination as on the sign of the transportation
-     */
-    public Tram(RequestQueue queue, @NonNull String lineId, @NonNull String direction, @Nullable String entityId, @NonNull Map<Station, DepartureTime> targetDestination) {
-        super(queue, lineId, direction, entityId, targetDestination);
-    }
-
-    @Override
-    public int getIconDrawableResource() {
-        return R.drawable.circle;
-    }
-}
